requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
redis[hiredis]==5.0.4
celery[redis]==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-multipart==0.0.9
alembic==1.13.1
psycopg2-binary==2.9.9
email-validator==2.1.1
